pipeline/news-pipeline/producer/requirements.txt
kafka-python==2.0.2
feedparser==6.0.11
beautifulsoup4==4.13.3
selenium
requests
python-dotenv
pydantic==1.10.15
loguru
